Move LiveTvConfigurationFactory to Configuration folder
